web/requirements.txt
django==3.2.5
django-markdown-deux==1.0.5